==> conflict_probability/01_notebooks/1_eth_conflict_master.ipynb <==
# Ethiopia conflict master data template
### ACLED: https://acleddata.com/data-export-tool/

import pandas as pd
import numpy as np
import os

url = 'https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/5c99f47f-ae3c-4935-98cf-72b310dfa50f/download/acled-1900-01-01-2020-06-04-eastern_africa.csv'
df = pd.read_csv(url, index_col=0)
list(df.columns)


df['month'] = pd.DatetimeIndex(df['event_date']).month

df = df[['country','admin1', 'admin2', 'year', 'month', 'fatalities', 'interaction']]

df['conflict']=1
df['interaction']=df['interaction'].astype(str)
df['cc_onset']=df['interaction'].str.contains('4')
df['actor_state']=df['interaction'].str.contains('14')
df['actor_political_militias']=df['interaction'].str.contains('34')
df['actor_identity_militias']=df['interaction'].str.contains('44')
df['actor_others']=df['interaction'].str.contains('48')
df['actor_rebel_groups']=df['interaction'].str.contains('24')
df['actor_civilians']=df['interaction'].str.contains('47')

df['cc_onset']=df['cc_onset'].astype(int)
df['actor_state']=df['actor_state'].astype(int)
df['actor_political_militias']=df['actor_political_militias'].astype(int)
df['actor_identity_militias']=df['actor_identity_militias'].astype(int)
df['actor_others']=df['actor_others'].astype(int)
df['actor_rebel_groups']=df['actor_rebel_groups'].astype(int)
df['actor_civilians']=df['actor_civilians'].astype(int)

df.columns

df = df.drop('interaction', 1)
df.columns

fatalities=df.groupby(['country','admin1', 'admin2', 'year', 'month']).sum().reset_index()

fatalities['cc_onset']=(fatalities['cc_onset'] > 0).astype(int)
fatalities['actor_state']=(fatalities['actor_state'] > 0).astype(int)
fatalities['actor_political_militias']=(fatalities['actor_political_militias'] > 0).astype(int)
fatalities['actor_identity_militias']=(fatalities['actor_identity_militias'] > 0).astype(int)
fatalities['actor_others']=(fatalities['actor_others'] > 0).astype(int)
fatalities['actor_rebel_groups']=(fatalities['actor_rebel_groups'] > 0).astype(int)
fatalities['actor_civilians']=(fatalities['actor_civilians'] > 0).astype(int)


fatalities = fatalities.set_index(['country', 'admin1','admin2','year', 'month'])
fatalities=fatalities.unstack().fillna(0).stack()
fatalities=fatalities.unstack().fillna(0).stack()
fatalities=pd.DataFrame(data=fatalities)
fatalities=fatalities.reset_index()

fatalities = fatalities.set_index(['country', 'admin1','admin2','month', 'year'])
fatalities=fatalities.unstack().fillna(0).stack()
fatalities=fatalities.unstack().fillna(0).stack()
fatalities=pd.DataFrame(data=fatalities)
fatalities=fatalities.reset_index()
fatalities

ea_conflict = fatalities[['country', 'admin1','admin2','year', 'month', 'fatalities', 'conflict', 'cc_onset', 'actor_state',
       'actor_political_militias', 'actor_identity_militias', 'actor_others',
       'actor_rebel_groups', 'actor_civilians']]

eth_conflict = ea_conflict.loc[ea_conflict['country']=='Ethiopia']

eth_conflict

eth_conflict.to_csv('/Users/yaredhurisa/Google Drive/Probability of conflict/indicators/eth_conflict_dataset_template.csv')

##### Actor code under inter1, inter2 and interaction
###### 1 = Governments and State Security Services
###### 2 = Rebel Groups
###### 3 = Political Militias
###### 4 = Identity Militias
###### 5 = Rioters
###### 6 = Protesters
###### 7 = Civilians
###### 8 = External/other forces


eth_conflict.shape
