==> conflict_probability/01_notebooks/2_merge_communal_conflict-eth.ipynb <==
# Meging comunal conflict dataset for Ethiopia
## Yared Hurisa
### 1. Conflict event data


import pandas as pd
import numpy as np

url = 'https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/02aa78d9-a90f-4249-8bb4-0c1147b5916d/download/eth_conflict_dataset_template.csv'
master = pd.read_csv(url, index_col=0)
list(master.columns)

print(master.shape)
master=master.drop_duplicates()
print(master.shape)

master = master.drop('country', 1)
master.columns

### 2. Inflation


url = 'https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/a4cfdfb8-f17f-4c7b-a350-fa1611cff5b0/download/eth_inflation.csv'
infl = pd.read_csv(url, index_col=0)

print(infl.shape)
infl=infl.drop_duplicates()
print(infl.shape)

df1=pd.merge(master, infl, how='left', on=['year', 'month'])
df1.shape

### 3. Rainfall

url = "https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/92077aa6-edf2-4533-afc4-837959da4594/download/eth_rainfall.csv"
rfl=pd.read_csv(url, index_col=0)

rfl.columns

rfl=rfl.rename(columns={'mean rainfall': 'rainfall'})

print(rfl.shape)
rfl=rfl.drop_duplicates()
print(rfl.shape)

rfl=rfl.replace({'admin2':{"Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
}})

rfl=rfl.replace({'admin1':{"Gambella":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benishangul Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia"
                         
}})

rfl=rfl.groupby(['admin1', 'admin2', 'year', 'month',])['rainfall'].mean().reset_index()
rfl.head()

df2=pd.merge(df1, rfl, how='left', on=['admin1', 'admin2', 'year', 'month'])
df2.shape

### 4. Temperature

url = "https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/8e54320f-8ef5-46ba-be58-f9fa5b3a1012/download/eth_temperature.csv"
temp=pd.read_csv(url, index_col=0)
temp.columns

print(temp.shape)
temp=temp.drop_duplicates()
print(temp.shape)

temp=temp.replace({'admin2':{"Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
}})

temp=temp.replace({'admin1':{"Gambella":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benishangul Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia"
                         
}})

temp=temp.groupby(['admin1', 'admin2', 'year', 'month',])['temperature mean'].mean().reset_index()
temp.head()


df3=pd.merge(df2, temp, how='left', on=['admin1', 'admin2', 'month', 'year'])

df3.shape

### 5. Population

url="https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/fcf0ade4-eab0-4cd2-ac04-7cc65ea67ccf/download/eth_pop.csv"
pop = pd.read_csv(url, index_col=0)
pop.columns

print(pop.shape)
pop=pop.drop_duplicates()
print(pop.shape)

pop=pop.replace({'admin2':{"Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                           "Wag Himra":"Wag Hemira",
                           "Bahir Dar":"Bar Dar Sp. Zone",
                           "Zone 03":"Zone 3",
                           "Zone 02":"Zone 2",
                           "Welwel & Warder":"Warder"
}})

pop=pop.replace({'admin1':{"Gambella":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benishangul Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia"
                         
}})

from fuzzywuzzy import process
clist=df3[['admin1']]
cdata=pop[['admin1']]
clist = clist.sort_values('admin1').drop_duplicates()
cdata = cdata.sort_values('admin1').drop_duplicates()

# create a choice list
choices = clist['admin1'].values.tolist()

# apply fuzzywuzzy to each row using lambda expression
cdata['Close Country'] = cdata['admin1'].apply(lambda x: process.extractOne(x, choices))

# add percent match wiht apply
cdata[['Close Country', 'Percent Match']] = cdata['Close Country'].apply(pd.Series)

cdata = cdata.sort_values('Percent Match')
cdata

from fuzzywuzzy import process
clist=df3[['admin2']]
cdata=pop[['admin2']]
clist = clist.sort_values('admin2').drop_duplicates()
cdata = cdata.sort_values('admin2').drop_duplicates()

# create a choice list
choices = clist['admin2'].values.tolist()

# apply fuzzywuzzy to each row using lambda expression
cdata['Close Country'] = cdata['admin2'].apply(lambda x: process.extractOne(x, choices))

# add percent match wiht apply
cdata[['Close Country', 'Percent Match']] = cdata['Close Country'].apply(pd.Series)

cdata = cdata.sort_values('Percent Match')
cdata = cdata[cdata['Percent Match']>70]
cdata

pop=pop.groupby(['admin1', 'admin2', 'year',])['btotl'].sum().reset_index()
pop.head()

df4=pd.merge(df3, pop, how='left', on=['admin1', 'admin2', 'year'])
df4.shape

# 6. Cereal Prices 

url="https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/84a11c8e-d2d1-4358-a33f-032cf6cb0d5a/download/eth_commodity_price.csv"
cereal = pd.read_csv(url, index_col=0)
cereal.columns

print(cereal.shape)
cereal=cereal.drop_duplicates()
print(cereal.shape)

cereal.isna().sum()

from fuzzywuzzy import process
clist=df4[['admin1']]
cdata=cereal[['admin1']]
clist = clist.sort_values('admin1').drop_duplicates()
cdata = cdata.sort_values('admin1').drop_duplicates()

# create a choice list
choices = clist['admin1'].values.tolist()

# apply fuzzywuzzy to each row using lambda expression
cdata['Close Country'] = cdata['admin1'].apply(lambda x: process.extractOne(x, choices))

# add percent match wiht apply
cdata[['Close Country', 'Percent Match']] = cdata['Close Country'].apply(pd.Series)

cdata = cdata.sort_values('Percent Match')
cdata

cereal=cereal.replace({'admin1':{"Gambella":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benishangul Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia"
                         
}})

from fuzzywuzzy import process
clist=df4[['admin2']]
cdata=cereal[['admin2']]
clist = clist.sort_values('admin2').drop_duplicates()
cdata = cdata.sort_values('admin2').drop_duplicates()

# create a choice list
choices = clist['admin2'].values.tolist()

# apply fuzzywuzzy to each row using lambda expression
cdata['Close Country'] = cdata['admin2'].apply(lambda x: process.extractOne(x, choices))

# add percent match wiht apply
cdata[['Close Country', 'Percent Match']] = cdata['Close Country'].apply(pd.Series)

cdata = cdata.sort_values('Percent Match')
cdata = cdata[cdata['Percent Match']>40]
cdata

cereal=cereal.replace({'admin2':{"Akaki Kaliti":"Addis Ababa",
                           "Nefas Silk Lafto":"Addis Ababa",
                           "LIDETA":"Addis Ababa",
                           "Kolfe Keraniyo":"Aaddis Ababa",
                           "Gulele":"Addis Ababa",
                           "Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                           "Wag Himra":"Wag Hemira",
                           "Bahir Dar":"Bar Dar Sp. Zone",
                           "Zone 03":"Zone 3",
                           "Zone 02":"Zone 2",
                           "Welwel & Warder":"Warder"
}})

d = {'January':1, 'February':2, 'March':3, 'April':4,'May':5, 'June':6, 'July':7, 'August':8,
     'September':9, 'October':10, 'November':11, 'December':12,}
cereal.month = cereal.month.map(d)
cereal

cereal=cereal.groupby(['admin1', 'admin2', 'year','month'])['maize_ETB_KG', 'teff_ETB_KG','wheat_ETB_KG'].mean().reset_index()
cereal.head()

df5=pd.merge(df4, cereal, how='left', on=['admin1', 'admin2', 'month', 'year'])

df5.shape

df5.columns


# 6. NDVI

url="https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/73f9adda-f5b9-4ee5-8932-39aaf53a2a82/download/eth_ndvi.csv"
ndvi = pd.read_csv(url, index_col=0)
ndvi.columns

print(ndvi.shape)
ndvi=ndvi.drop_duplicates()
print(ndvi.shape)

ndvi=ndvi.replace({'admin1':{"Gambella":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benishangul Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia"
                         
}})

ndvi=ndvi.replace({'admin2':{"Akaki Kaliti":"Addis Ababa",
                           "Nefas Silk Lafto":"Addis Ababa",
                           "LIDETA":"Addis Ababa",
                           "Kolfe Keraniyo":"Aaddis Ababa",
                           "Gulele":"Addis Ababa",
                           "Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                           "Wag Himra":"Wag Hemira",
                           "Bahir Dar":"Bar Dar Sp. Zone",
                           "Zone 03":"Zone 3",
                           "Zone 02":"Zone 2",
                           "Welwel & Warder":"Warder"
}})

ndvi=ndvi.groupby(['admin1', 'admin2', 'year','month'])['mean ndvi', 'ndvi_lag'].mean().reset_index()
ndvi.head()

df6=pd.merge(df5, ndvi, how='left', on=['admin1', 'admin2', 'year', 'month'])

df6.shape

df6['group']=df6['admin1'].astype(str)+'_'+df6['admin2'].astype(str)
df6.columns

## Food security classification

url=('https://data.kimetrica.com/dataset/4dbc3cc7-9474-49f2-bfd4-231e78401caa/resource/0456cb8c-274a-4001-9089-8df2533dbd3b/download/eth_food_security_classification.csv')
fs = pd.read_csv(url, index_col=0)
fs.columns

fs=fs[['year', 'month', 'admin1', 'admin2', 'fs_Crisis',
       'fs_Emergency', 'fs_Famine', 'fs_Minimal', 'fs_Stressed']]

fs=fs.replace({'admin1':{"Gambela":"Gambela Peoples",
                         "Harari":"Harari People",
                         "Benshangul-Gumuz":"Benshangul-Gumaz",
                         "Oromiya":"Oromia",
                         "SNNPR":"Southern Nations"
                         
}})

fs=fs.replace({'admin2':{"Akaki Kaliti":"Addis Ababa",
                           "Nefas Silk Lafto":"Addis Ababa",
                           "LIDETA":"Addis Ababa",
                           "Kolfe Keraniyo":"Aaddis Ababa",
                           "Gulele":"Addis Ababa",
                           "Ilu Aba Bora":"Illubabor",
                         "Bahir Dar":"Bar Dar Sp. Zone",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                         "East Hararge":"East Harerghe",
                         "West Hararge":"West Harerghe",
                         "South Tigray":"Southern Tigray",
                         "Nuwer":"Nuer",
                         "Kefa":"Keffa",
                         "West Arsi":"Arsi",
                         "Jimma Sp.":"Jimma",
                         "Siliti":"Silti",
                         "Assosa":"Asosa",
                         "Dawuro":"Dawro",
                         "Zone 01":"Zone 1",
                         "Agnuwak":"Agnuak",
                         "Zone 04":"Zone 4",
                         "Kembata Timbaro":"Kembata Tembaro",
                         "West Gojjam":"West Gojam",
                         "North Western Tigray":"Western Tigray",
                         "South Wolo":"South Wollo",
                         "East Gojjam":"East Gojam",
                         "Horo Gudru Wellega":"Horo Gudru Welega",
                         "Zone 03":"Zone 3",
                         "Zone 02":"Zone 2",
                           "Wag Himra":"Wag Hemira",
                           "Bahir Dar":"Bar Dar Sp. Zone",
                           "Zone 03":"Zone 3",
                           "Zone 02":"Zone 2",
                           "Welwel & Warder":"Warder",
                         "Agniwak":"Agnuak",
                         "Kemashi":"Kamashi",
                         "Horo Guduru":"Horo Gudru Welega",
                         "Shebelle":"Shabelle",
                         "East Tigray":"Eastern Tigray",
                         "West Tigray":"Western Tigray",
                         "Gujii":"Guji",
                         "Ilubabor":"Illubabor"
                         
                         
}})

fs=fs.groupby(['admin1', 'admin2', 'year','month'])['fs_Crisis', 'fs_Emergency','fs_Famine', 'fs_Minimal', 'fs_Stressed'].sum().reset_index()
fs.head()

df7=pd.merge(df6, fs, how='left', on=['admin1', 'admin2', 'year', 'month'])

print(df7.shape)
print(df7.isna().sum())

list(df7.columns)

# Missing data handling using MICE


impute_list=[
'inflation_all',
 'inflation_food',
 'inflation_non_food',
 'rainfall',
 'temperature mean',
 'btotl',
 'maize_ETB_KG',
 'teff_ETB_KG',
 'wheat_ETB_KG',
 'mean ndvi',
 'ndvi_lag',]


# impute using MICE method
from impyute.imputation import cs
mice_impute=cs.mice(df7[impute_list].values)
df_mice=pd.DataFrame(mice_impute, columns=df7[impute_list].columns, 
                     index=df7.index)

print(df_mice.isna().sum())

df7[impute_list]=df_mice
print(df7.isna().sum())

df7=df7.fillna(0)

df7 = df7.drop('admin0', 1)

df7.to_csv('/Users/yaredhurisa/Google Drive/Probability of conflict/indicators/dataset_communal_cnflict_model.csv')
